+++ harvey/config.py
@@ -1,0 +1,44 @@
+import os
+
+from dotenv import load_dotenv
+
+from version import VERSION
+
+load_dotenv()  # Must remain at the top of this file
+
+
+class Config:
+    # User configurable settings
+    allowed_branches = [branch.strip().lower() for branch in os.getenv('ALLOWED_BRANCHES', 'main,master').split(',')]
+    deploy_timeout = int(os.getenv('DEPLOY_TIMEOUT', 1800))  # Default is 30 minutes
+    git_timeout = int(os.getenv('GIT_TIMEOUT', 300))  # Default is 5 minutes
+    pagination_limit = int(os.getenv('PAGINATION_LIMIT', 20))
+    deploy_on_tag = os.getenv('DEPLOY_ON_TAG', True)  # Whether a tag pushed will trigger a deploy or not
+    use_slack = os.getenv('USE_SLACK')
+    webhook_secret = os.getenv('WEBHOOK_SECRET', '')
+    slack_bot_token = os.getenv('SLACK_BOT_TOKEN')
+    slack_channel = os.getenv('SLACK_CHANNEL', 'general')
+    harvey_path = os.getenv('HARVEY_PATH', os.path.expanduser('~/harvey'))
+
+    # Harvey settings
+    host = os.getenv('HOST', '127.0.0.1')
+    port = os.getenv('PORT', '5000')
+    harvey_version = VERSION
+    supported_pipelines = {
+        'deploy',
+        'pull',
+    }
+    projects_path = os.path.join(harvey_path, 'projects')
+    stores_path = os.path.join(harvey_path, 'stores')
+    pipelines_store_path = os.path.join(stores_path, 'pipelines.sqlite')
+    locks_store_path = os.path.join(stores_path, 'locks.sqlite')
+    logger_name = 'harvey'
+    log_level = os.getenv('LOG_LEVEL', 'INFO').upper()
+    log_location = os.path.join(harvey_path, 'logs')
+
+    # Emoji (used for Slack messages, set defaults if slack isn't in use)
+    # TODO: Defaults are nice for when slack isn't in use; however, the emoji text will
+    # still show up in log files, we should be writting the fallback message to logs instead of emojis
+    work_emoji = ':hammer_and_wrench:' if use_slack else ''
+    success_emoji = ':white_check_mark:' if use_slack else 'Success!'
+    failure_emoji = ':skull_and_crossbones:' if use_slack else 'Failure!'