+++ harvey/git.py
@@ -4,15 +4,16 @@ from typing import Any, Dict, List
 
 import woodchips
 
-from harvey.globals import Global
-from harvey.utils import LOGGER_NAME, Utils
+from harvey.config import Config
+from harvey.utils import Utils
+from harvey.webhooks import Webhook
 
 
 class Git:
     @staticmethod
     def update_git_repo(webhook: Dict[str, Any]) -> str:
         """Clone or pull repo using Git depending on if it exists or not."""
-        project_path = os.path.join(Global.PROJECTS_PATH, Global.repo_full_name(webhook))
+        project_path = os.path.join(Config.projects_path, Webhook.repo_full_name(webhook))
 
         if os.path.exists(project_path):
             output = Git.pull_repo(project_path, webhook)
@@ -24,11 +25,11 @@ class Git:
     @staticmethod
     def clone_repo(project_path: str, webhook: Dict[str, Any]) -> str:
         """Clone a repo into the Harvey projects folder."""
-        logger = woodchips.get(LOGGER_NAME)
+        logger = woodchips.get(Config.logger_name)
         decoded_output = ''
 
         try:
-            command = ['git', 'clone', '--depth=1', Global.repo_url(webhook), project_path]
+            command = ['git', 'clone', '--depth=1', Webhook.repo_url(webhook), project_path]
             command_output = Git._git_subprocess(command)
             decoded_output = command_output.decode()
             logger.debug(decoded_output)
@@ -37,7 +38,7 @@ class Git:
             logger.error(final_output)
             Utils.kill(final_output, webhook)
         except subprocess.CalledProcessError:
-            final_output = f'Harvey could not clone {Global.repo_full_name(webhook)}.'
+            final_output = f'Harvey could not clone {Webhook.repo_full_name(webhook)}.'
             logger.error(final_output)
             Utils.kill(final_output, webhook)
 
@@ -46,7 +47,7 @@ class Git:
     @staticmethod
     def pull_repo(project_path: str, webhook: Dict[str, Any], pull_attempt: int = 1) -> str:
         """Pull updates for a repo in the Harvey projects folder."""
-        logger = woodchips.get(LOGGER_NAME)
+        logger = woodchips.get(Config.logger_name)
         decoded_output = ''
 
         try:
@@ -61,8 +62,8 @@ class Git:
         except subprocess.CalledProcessError:
             # The biggest offender of this action failing is local, uncommitted changes.
             # Try stashing and retry again before failing.
-            logger.error(f'Harvey could not pull {Global.repo_full_name(webhook)}.')
-            logger.info(f'Attempting to stash {Global.repo_full_name(webhook)} before pulling again...')
+            logger.error(f'Harvey could not pull {Webhook.repo_full_name(webhook)}.')
+            logger.info(f'Attempting to stash {Webhook.repo_full_name(webhook)} before pulling again...')
 
             try:
                 command = ['git', '-C', project_path, 'stash']
@@ -70,7 +71,7 @@ class Git:
                 decoded_output = command_output.decode()
                 logger.debug(f'{decoded_output}')
             except (subprocess.CalledProcessError, subprocess.TimeoutExpired):
-                final_output = f'Harvey could not stash {Global.repo_full_name(webhook)}.'
+                final_output = f'Harvey could not stash {Webhook.repo_full_name(webhook)}.'
                 logger.error(final_output)
                 Utils.kill(final_output, webhook)
 
@@ -88,7 +89,7 @@ class Git:
             command,
             stdin=None,
             stderr=None,
-            timeout=Global.GIT_TIMEOUT,
+            timeout=Config.git_timeout,
         )
 
         return command_output