+++ harvey/utils.py
@@ -1,17 +1,13 @@
 import datetime
-import os
 import sys
 from typing import Any, Dict
 
 import woodchips
 from sqlitedict import SqliteDict  # type: ignore
 
-from harvey.globals import Global
+from harvey.config import Config
 from harvey.messages import Message
-
-LOG_LEVEL = os.getenv('LOG_LEVEL', 'INFO').upper()
-LOGGER_NAME = 'harvey'
-LOG_LOCATION = os.path.join(Global.HARVEY_PATH, 'logs')
+from harvey.webhooks import Webhook
 
 
 class Utils:
@@ -20,15 +16,15 @@ class Utils:
         """A kill util to write everything to logs, send messages,
         tear down Docker stuff, and quit.
         """
-        logger = woodchips.get(LOGGER_NAME)
+        logger = woodchips.get(Config.logger_name)
 
-        failure_message = f'{Global.repo_full_name(webhook)} pipeline failed!'
+        failure_message = f'{Webhook.repo_full_name(webhook)} pipeline failed!'
         logger.warning(failure_message)
 
         pipeline_logs = final_output + failure_message
         Utils.store_pipeline_details(webhook, pipeline_logs)
 
-        if Global.SLACK:
+        if Config.use_slack:
             Message.send_slack_message(pipeline_logs)
 
         Utils.update_project_lock(webhook=webhook, locked=False)
@@ -39,15 +35,15 @@ class Utils:
     @staticmethod
     def success(final_output: str, webhook: Dict[str, Any]):
         """Log output and send message on pipeline success."""
-        logger = woodchips.get(LOGGER_NAME)
+        logger = woodchips.get(Config.logger_name)
 
-        success_message = f'{Global.repo_full_name(webhook)} pipeline succeeded!'
+        success_message = f'{Webhook.repo_full_name(webhook)} pipeline succeeded!'
         logger.info(success_message)
 
         pipeline_logs = final_output + success_message
         Utils.store_pipeline_details(webhook, pipeline_logs)
 
-        if Global.SLACK:
+        if Config.use_slack:
             Message.send_slack_message(pipeline_logs)
 
         Utils.update_project_lock(webhook=webhook, locked=False)
@@ -61,11 +57,11 @@ class Utils:
 
         A project ID consists of the `project_name@commit_id`.
         """
-        logger = woodchips.get(LOGGER_NAME)
+        logger = woodchips.get(Config.logger_name)
 
-        logger.debug(f'Storing pipeline details for {Global.repo_full_name(webhook)}...')
+        logger.debug(f'Storing pipeline details for {Webhook.repo_full_name(webhook)}...')
 
-        with SqliteDict(Global.PIPELINES_STORE_PATH) as mydict:
+        with SqliteDict(Config.pipelines_store_path) as mydict:
             # Naively check the logs for an indicator of the status being success
             if 'success' in final_output.lower() or 'succeeded' in final_output.lower():
                 pipeline_status = 'Success'
@@ -74,9 +70,9 @@ class Utils:
             else:
                 pipeline_status = 'Failure'
 
-            mydict[Global.pipeline_id(webhook)] = {
-                'project': Global.repo_full_name(webhook).replace("/", "-"),
-                'commit': Global.repo_commit_id(webhook),
+            mydict[Webhook.pipeline_id(webhook)] = {
+                'project': Webhook.repo_full_name(webhook).replace("/", "-"),
+                'commit': Webhook.repo_commit_id(webhook),
                 'log': final_output,
                 'status': pipeline_status,
                 'timestamp': str(datetime.datetime.utcnow()),
@@ -91,13 +87,13 @@ class Utils:
         Locking should only happen once a pipeline is begun. A locked deployment should then always be
         unlocked once it's finished regardless of status so another deployment can follow.
         """
-        logger = woodchips.get(LOGGER_NAME)
+        logger = woodchips.get(Config.logger_name)
 
         locked_string = 'Locking' if locked is True else 'Unlocking'
-        logger.info(f'{locked_string} deployments for {Global.repo_full_name(webhook)}...')
+        logger.info(f'{locked_string} deployments for {Webhook.repo_full_name(webhook)}...')
 
-        with SqliteDict(Global.LOCKS_STORE_PATH) as mydict:
-            mydict[Global.repo_full_name(webhook).replace("/", "-")] = {
+        with SqliteDict(Config.locks_store_path) as mydict:
+            mydict[Webhook.repo_full_name(webhook).replace("/", "-")] = {
                 'locked': locked,
             }
 
@@ -109,7 +105,7 @@ class Utils:
         locked_value = False
         corrected_project_name = project_name.replace("/", "-")
 
-        with SqliteDict(Global.LOCKS_STORE_PATH) as mydict:
+        with SqliteDict(Config.locks_store_path) as mydict:
             for key, value in mydict.iteritems():
                 if key == corrected_project_name:
                     locked_value = value['locked']
@@ -121,8 +117,8 @@ class Utils:
 def setup_logger():
     """Sets up a `woodchips` logger instance."""
     logger = woodchips.Logger(
-        name=LOGGER_NAME,
-        level=LOG_LEVEL,
+        name=Config.logger_name,
+        level=Config.log_level,
     )
     logger.log_to_console()
-    logger.log_to_file(location=LOG_LOCATION)
+    logger.log_to_file(location=Config.log_location)