+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## NEXT RELEASE
 
+* Overhauled the pipeline logs to provide more details that are easier to read
 * Changed the required `harvey.json` file for each project to `.harvey.yml` (expected fields/values remain the same)
 * Sets the default pagination limit down to 20 from 100, adds `PAGINATION_LIMIT` env var to allow customization
 * Added new `HARVEY_PATH`, `GIT_TIMEOUT`, `DEPLPOY_TIMEOUT`, and `DEPLOY_ON_TAG` env vars for customization
 * Better syncing of the version string for the release and what appears in logs
 * Better error handling around `404`s at the API level
+* Various bug fixes and refactoring improvements
 
 ## v0.16.0 (2022-01-18)